Data is now responsible for ball movement.
