--- Data/BallChangedEventArgs.cs
@@ -1,11 +1,0 @@
-﻿namespace BallSimulator.Data;
-
-public class BallChangedEventArgs : EventArgs
-{
-    public IBallDto Ball { get; set; }
-
-    public BallChangedEventArgs(IBallDto ball)
-    {
-        Ball = ball;
-    }
-}
