--- Logic/BallChangedEventArgs.cs
@@ -1,11 +1,0 @@
-﻿namespace BallSimulator.Logic;
-
-public class BallChangedEventArgs : EventArgs
-{
-    public IBall Ball { get; init; }
-
-    public BallChangedEventArgs(IBall ball)
-    {
-        Ball = ball;
-    }
-}
