+++ Model/IBallModel.cs
@@ -3,7 +3,7 @@ using System.ComponentModel;
 
 namespace BallSimulator.Presentation.Model;
 
-public interface IBallModel : IObserver<IBall>, INotifyPropertyChanged
+public interface IBallModel : IObserver<IBallLogic>, INotifyPropertyChanged
 {
     public int Diameter { get; }
     public int Radius { get; }
