+++ Model/ModelApi.cs
@@ -6,7 +6,7 @@ internal class ModelApi : ModelAbstractApi
 {
     private readonly LogicAbstractApi _logic;
     private readonly ISet<IObserver<IBallModel>> _observers;
-    private readonly IDictionary<IBall, IBallModel> _ballToBallModel;
+    private readonly IDictionary<IBallLogic, IBallModel> _ballToBallModel;
 
     private IDisposable? _unsubscriber;
 
@@ -14,7 +14,7 @@ internal class ModelApi : ModelAbstractApi
     {
         _logic = logic ?? LogicAbstractApi.CreateLogicApi();
         _observers = new HashSet<IObserver<IBallModel>>();
-        _ballToBallModel = new Dictionary<IBall, IBallModel>();
+        _ballToBallModel = new Dictionary<IBallLogic, IBallModel>();
     }
 
     public override void Start(int ballsCount)
@@ -30,7 +30,7 @@ internal class ModelApi : ModelAbstractApi
 
     #region Observer
 
-    public void Follow(IObservable<IBall> provider)
+    public void Follow(IObservable<IBallLogic> provider)
     {
         _unsubscriber = provider.Subscribe(this);
     }
@@ -41,7 +41,7 @@ internal class ModelApi : ModelAbstractApi
         EndTransmission();
     }
 
-    public override void OnNext(IBall ball)
+    public override void OnNext(IBallLogic ball)
     {
         _ballToBallModel.TryGetValue(ball, out var ballModel);
         if (ballModel is null)
